Fixed issues & added draft bookings
