+++ lib/screens/profile_screen.dart
@@ -1,7 +1,7 @@
 import 'package:base42_events_mobile/nav.dart';
 import 'package:base42_events_mobile/providers/auth_provider.dart';
 import 'package:base42_events_mobile/theme.dart';
-import 'package:base42_events_mobile/widgets/booking_card.dart';
+import 'package:base42_events_mobile/widgets/booking/booking_card.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:go_router/go_router.dart';
