+++ lib/types.dart
@@ -16,3 +16,57 @@ class AuthResponse {
     );
   }
 }
+
+class BookingRequestPayload {
+  final String organizerEntity;
+  final String initiatorName;
+  final String email;
+  final String phone;
+  final String companyName;
+  final String eventType;
+  final String eventName;
+  final String eventTheme;
+  final String eventPurpose;
+  final String eventAgenda;
+  final String eventDate;
+  final String eventStartTime;
+  final String eventEndTime;
+  final String physicalPresence;
+  final String expectedGuests;
+
+  const BookingRequestPayload({
+    required this.organizerEntity,
+    required this.initiatorName,
+    required this.email,
+    required this.phone,
+    required this.companyName,
+    required this.eventType,
+    required this.eventName,
+    required this.eventTheme,
+    required this.eventPurpose,
+    required this.eventAgenda,
+    required this.eventDate,
+    required this.eventStartTime,
+    required this.eventEndTime,
+    required this.physicalPresence,
+    required this.expectedGuests,
+  });
+
+  Map<String, dynamic> toStrapiData() => {
+    'company-name': companyName,
+    'email': email,
+    'event-agenda': eventAgenda,
+    'event-date': eventDate,
+    'event-end-time': eventEndTime,
+    'event-name': eventName,
+    'event-purpose': eventPurpose,
+    'event-start-time': eventStartTime,
+    'event-theme': eventTheme,
+    'event-type': eventType,
+    'expected-guests': int.tryParse(expectedGuests.trim()) ?? 0,
+    'initiator-name': initiatorName,
+    'organizer-entity': organizerEntity,
+    'phone': phone,
+    'physical-presence': physicalPresence.trim().toLowerCase(),
+  };
+}
